+++ src/Razor/src/Microsoft.VisualStudio.Mac.LanguageServices.Razor/ProjectBuildChangeTrigger.cs
@@ -4,6 +4,7 @@
 using System;
 using System.ComponentModel.Composition;
 using System.Linq;
+using System.Threading;
 using Microsoft.CodeAnalysis.Razor;
 using Microsoft.CodeAnalysis.Razor.ProjectSystem;
 using Microsoft.VisualStudio.Editor.Razor;
@@ -127,7 +128,7 @@ namespace Microsoft.VisualStudio.Mac.LanguageServices.Razor
                 {
                     // Trigger a tag helper update by forcing the project manager to see the workspace Project
                     // from the current solution.
-                    _workspaceStateGenerator.Update(workspaceProject, projectSnapshot);
+                    _workspaceStateGenerator.Update(workspaceProject, projectSnapshot, CancellationToken.None);
                 }
             }
         }
