+++ src/Razor/test/Microsoft.VisualStudio.Editor.Razor.Test.Common/TestProjectWorkspaceStateGenerator.cs
@@ -24,7 +24,7 @@ namespace Microsoft.VisualStudio.Editor.Razor.Test
         {
         }
 
-        public override void Update(Project workspaceProject, ProjectSnapshot projectSnapshot, CancellationTokenSource cts = default)
+        public override void Update(Project workspaceProject, ProjectSnapshot projectSnapshot, CancellationToken cancellationToken)
         {
             var update = (workspaceProject, projectSnapshot);
             _updates.Add(update);
